# openapi.yaml
openapi: 3.0.0
info:
  title: Student Management System API
  version: 1.0.0
  description: |
    Comprehensive API documentation for managing students, institutes, courses, and academic results
servers:
  - url: http://localhost:3000/api
    description: Development server
components:
  parameters:
    studentIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    instituteIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    resultIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Student:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        instituteId:
          type: string
          format: uuid
        status:
          type: string
          enum: [active, inactive]
...
    Result:
      type: object
      properties:
        id:
          type: string
          format: uuid
        score:
          type: integer
          minimum: 0
          maximum: 100
        academicYear:
          type: integer
        semester:
          type: string
          enum: [Spring, Summer, Fall]
        studentId:
          type: string
          format: uuid
        courseId:
          type: string
          format: uuid
    Error:
      type: object
      properties:
        status_code:
          type: integer
        status:
          type: string
        message:
          type: string
        error:
          type: string
paths:
...
  /students:
    get:
      tags: [Students]
      summary: Get paginated students
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status_code:
                    type: integer
                  status:
                    type: string
                  data:
                    type: object
                    properties:
                      students:
                        type: array
                        items:
                          $ref: "#/components/schemas/Student"
                      pagination:
                        type: object
                        properties:
                          total:
                            type: integer
                          page:
                            type: integer
                          pages:
                            type: integer
    post:
      tags: [Students]
      summary: Create new student
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Student"
      responses:
        201:
          description: Student created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status_code:
                    type: integer
                  status:
                    type: string
                  data:
                    type: object
                    properties:
                      student:
                        $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      tags: [Students]
      summary: Get student by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/studentIdParam"
      responses:
        200:
          description: Student details
          content:
            application/json:
              schema:
                type: object
                properties:
                  status_code:
                    type: integer
                  status:
                    type: string
                  data:
                    type: object
                    properties:
                      student:
                        $ref: "#/components/schemas/Student"
        404:
          description: Student not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags: [Students]
      summary: Update student details
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/studentIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Student"
      responses:
        200:
          description: Student updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status_code:
                    type: integer
                  status:
                    type: string
                  data:
                    type: object
                    properties:
                      student:
                        $ref: "#/components/schemas/Student"
        404:
          description: Student not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Students]
      summary: Delete student
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/studentIdParam"
      responses:
        200:
          description: Student deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status_code:
                    type: integer
                  status:
                    type: string
                  message:
                    type: string
        404:
          description: Student not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /results:
    get:
      tags: [Results]
      summary: List all results
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status_code:
                    type: integer
                  status:
                    type: string
                  data:
                    type: object
                    properties:
                      results:
                        type: array
                        items:
                          $ref: "#/components/schemas/Result"
                      pagination:
                        type: object
                        properties:
                          total:
                            type: integer
                          page:
                            type: integer
                          pages:
                            type: integer
    post:
      tags: [Results]
      summary: Create new result
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Result"
      responses:
        201:
          description: Result created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status_code:
                    type: integer
                  status:
                    type: string
                  data:
                    type: object
                    properties:
                      result:
                        $ref: "#/components/schemas/Result"
  /results/{id}:
    get:
      tags: [Results]
      summary: Get result by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/resultIdParam"
      responses:
        200:
          description: Result details
          content:
            application/json:
              schema:
                type: object
                properties:
                  status_code:
                    type: integer
                  status:
                    type: string
                  data:
                    type: object
                    properties:
                      result:
                        $ref: "#/components/schemas/Result"
        404:
          description: Result not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags: [Results]
      summary: Update result
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/resultIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Result"
      responses:
        200:
          description: Result updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status_code:
                    type: integer
                  status:
                    type: string
                  data:
                    type: object
                    properties:
                      result:
                        $ref: "#/components/schemas/Result"
        404:
          description: Result not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Results]
      summary: Delete result
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/resultIdParam"
      responses:
        200:
          description: Result deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status_code:
                    type: integer
                  status:
                    type: string
                  message:
                    type: string
        404:
          description: Result not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /reports/students/{instituteId}:
    get:
      tags: [Reports]
      summary: Get students by institute
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/instituteIdParam"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        200:
          description: Students list by institute
          content:
            application/json:
              schema:
                type: object
                properties:
                  status_code:
                    type: integer
                  status:
                    type: string
                  data:
                    type: object
                    properties:
                      students:
                        type: array
                        items:
                          $ref: "#/components/schemas/Student"
                      pagination:
                        type: object
                        properties:
                          total:
                            type: integer
                          page:
                            type: integer
                          pages:
                            type: integer
...